add login/register + firebase auth + other code changes
