--- app/src/main/java/com/example/catchio/navigation/Screen.kt
@@ -1,7 +1,0 @@
-package com.example.catchio.navigation
-
-enum class Screen(val route: String) {
-    Main("main"),
-    Route("route"),
-    Backpack("backpack")
-}
