--- app/src/main/java/com/example/catchio/screens/MainScreen.kt
@@ -1,23 +1,0 @@
-package com.example.catchio.screens
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
-
-@Composable
-fun MainScreen(mainViewModel: MainViewModel = viewModel()) {
-    val state by mainViewModel.state.collectAsState()
-    Column(
-        modifier = Modifier.fillMaxSize().padding(16.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(state)
-    }
-}
