--- app/src/main/java/com/example/catchio/screens/MainViewModel.kt
@@ -1,10 +1,0 @@
-package com.example.catchio.screens
-
-import androidx.lifecycle.ViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-
-class MainViewModel : ViewModel() {
-    private val _state = MutableStateFlow("Main Screen Content")
-    val state: StateFlow<String> = _state
-}
