--- app/src/main/java/com/example/catchio/screens/town/Town.kt
@@ -1,11 +1,0 @@
-package com.example.catchio.screens.town
-
-data class Town(
-    val townData: TownData
-) {
-    data class TownData(
-        val dragonName: String? = null,
-        val levelRange: IntRange? = null,
-        val imageResId: Int? = null
-    )
-}
