+++ app/src/main/java/com/vcoffee/catchio/navigation/NavigationUtils.kt
@@ -1,0 +1,15 @@
+package com.vcoffee.catchio.navigation
+
+import androidx.navigation.NavGraph.Companion.findStartDestination
+import androidx.navigation.NavHostController
+
+fun NavHostController.navigateSingleTopTo(route: String) =
+    this.navigate(route) {
+        popUpTo(
+            this@navigateSingleTopTo.graph.findStartDestination().id
+        ) {
+            saveState = true
+        }
+        launchSingleTop = true
+        restoreState = true
+    }
