+++ app/src/main/java/com/vcoffee/catchio/navigation/Screen.kt
@@ -1,0 +1,7 @@
+package com.vcoffee.catchio.navigation
+
+enum class Screen(val route: String) {
+    Main("main"),
+    Route("route"),
+    Backpack("backpack")
+}
