+++ app/src/main/java/com/vcoffee/catchio/screens/route/RouteScreen.kt
@@ -1,0 +1,64 @@
+package com.vcoffee.catchio.screens.route
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavHostController
+
+@Composable
+fun Screen1(navController: NavHostController, routeScreenViewModel: RouteScreenViewModel = viewModel()) {
+    val rowsCount by routeScreenViewModel.rowsCount.collectAsState()
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Color.Green),
+        contentAlignment = Alignment.Center
+    ) {
+        LazyColumn(
+            modifier = Modifier.fillMaxSize(),
+            verticalArrangement = Arrangement.spacedBy(16.dp),
+            contentPadding = PaddingValues(16.dp)
+        ) {
+            items(rowsCount) { rowIndex ->
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.spacedBy(16.dp)
+                ) {
+                    for (colIndex in 0..1) {
+                        val isRed = (rowIndex + colIndex) % 2 == 0
+                        Box(
+                            modifier = Modifier
+                                .weight(1f)
+                                .aspectRatio(1f)
+                                .background(
+                                    color = if (isRed) Color.Red else Color.Transparent,
+                                    shape = RoundedCornerShape(8.dp)
+                                )
+                                .clickable(enabled = isRed) {
+                                    if (isRed) {
+                                        navController.navigate("townDetails/$rowIndex/$colIndex")
+                                    }
+                                }
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
