+++ app/src/main/java/com/vcoffee/catchio/screens/route/RouteScreenViewModel.kt
@@ -1,0 +1,10 @@
+package com.vcoffee.catchio.screens.route
+
+import androidx.lifecycle.ViewModel
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+class RouteScreenViewModel : ViewModel() {
+    private val _rowsCount = MutableStateFlow(10)
+    val rowsCount: StateFlow<Int> = _rowsCount
+}
